chore(servers): move loadout preview into own component
